User need to be authorized to delete a room. A user can delete its own messages

+++ routes/chatRoutes.js
@@ -28,7 +28,7 @@ router.get("/:id", isLoggedIn, async (req, res) => {
         return Promise.all(oldMessages)
     })
         .then(messageList => {
-            console.log(messageList)
+            // console.log(messageList)
             res.render("chatroom", { id, user, currentRoom, messageList })
         })
 })
@@ -43,13 +43,18 @@ router.post("/", isLoggedIn, async (req, res) => {
 
 //endpoint for deleting a room 
 //todo: id a room is deleted, all the messages in that room should also be deleted
+// want to make sure that the user is loggedin but also that the user is allowed to delete this chatroom (authorized)
 router.delete("/:id", isLoggedIn, async (req, res) => {
     const { id } = req.params;
-    console.log(id);
-    await Room.findByIdAndDelete(id)
-    res.end("room was deleted")
+    const room = await Room.findById(id).populate("creator")
+    if (!(JSON.stringify(req.user._id) == JSON.stringify(room.creator._id))) {
+        req.flash("error", "you do not have permission to delete this room")
+        return res.end("no room was deleted")
+    } else {
+        await Room.findByIdAndDelete(id)
+        req.flash("success", `${room.name} is now deleted`)
+        res.end("room deleted")
+    }
 })
-
-
 
 module.exports = router
